gcpProject/udemy-cloud-monkey-cloud-run-code-repo/lab-17-integration-with-cloud-storage-using-eventarc/microservices/customers/requirements.txt
blinker==1.6.2
cachetools==5.3.1
click==8.1.6
cloudevents==1.9.0
deprecation==2.1.0
Flask==2.3.2
functions-framework==3.4.0
google-auth==2.17.3
google-events==0.10.0
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.1.0
packaging==23.1
proto-plus==1.22.3
protobuf==4.21.12
pyasn1==0.5.0
pyasn1-modules==0.3.0
rsa==4.9
six==1.16.0
watchdog==3.0.0
Werkzeug==2.3.6
google-cloud-storage==2.10.0